refactor: factor out parseYAML, parseJSON, state mutation logic
